Order guesses by the widest bucket
Reuse some of the earlier work of ordering by the most convenient
buckets. This means that we're trying the guess with the smallest
maximum bucket size first; or as previously defined, the best guess.

Use a microoptimization of precompute the scores and use set
intersection to compute this quickly, about a 6x speedup over computing
the buckets.

Seems to narrow the tree a good bit faster, but who's to say